Brake computing and timeout changes
The braking angle is not starting from 0 but from the current angle
given by the sensor. This avoid oscillating if there is angle overshoot
after sharp release of stick from big angle.
Timeout is computed differently and avoid errors that could be due to
the wind effect or the overshoot mentionned before. Moreover, in the
last timeout code , timeout could never be reached.